+++ addon/components/ivy-tabs-tablist.js
@@ -30,6 +30,16 @@ export default Component.extend({
     this.get('tabsContainer').unregisterTabList(this);
   },
 
+
+  /**
+   * The label of the tablist for screenreaders to use.
+   *
+   * @property aria-label
+   * @type String
+   * @default ''
+   */
+  'aria-label': '',
+
   /**
    * Tells screenreaders to notify the user during DOM modifications.
    *
@@ -77,7 +87,7 @@ export default Component.extend({
     }
   }).readOnly(),
 
-  attributeBindings: ['aria-live', 'aria-multiselectable', 'aria-relevant'],
+  attributeBindings: ['aria-label', 'aria-live', 'aria-multiselectable', 'aria-relevant'],
 
   classNames: ['ivy-tabs-tablist'],